Extract matcher-algo-specific logic
